dim_predictor/requirements.txt
gradio
ultralytics>=8.0.100
opencv-python-headless==4.8.1.78
numpy
pillow
torch
torchvision